# specs/001-unified-chatbot-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Maliev Unified Chatbot Service
  version: 1.0.0
  description: API for the unified conversational intelligence layer of the MALIEV ecosystem.
servers:
  - url: /chatbot
    description: Service prefix
paths:
...
  /v1/messages:
    post:
      summary: Send a message to the chatbot (Website/Intranet channel)
      operationId: sendMessage
      tags: [Messages]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '429':
          description: Rate limit exceeded (100 msg/hr)
...
components:
  schemas:
...
    SendMessageRequest:
      type: object
      required: [sessionId, content]
      properties:
        sessionId:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 10000
        attachments:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/Attachment'
    MessageResponse:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
        content:
          type: string
        role:
          type: string
          enum: [assistant]
        language:
          type: string
          enum: [en, th]
        suggestedActions:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedAction'
        processingTimeMs:
          type: integer
    Attachment:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [image, audio, pdf, video]
        url:
          type: string
          format: uri
        sizeBytes:
          type: integer
          maximum: 52428800  # 50MB
    SuggestedAction:
      type: object
      properties:
        label:
          type: string
        action:
          type: string
          enum: [reply, link, api_call]
        payload:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with IAM permissions (chatbot.instructions.read, chatbot.instructions.update)